Struct sui_framework_build::compiled_package::CompiledPackage

pub struct CompiledPackage {
    pub package: MoveCompiledPackage,
    pub path: PathBuf,
}

Wrapper around the core Move CompiledPackage with some Sui-specific traits and info

Fields
package: MoveCompiledPackagepath: PathBuf

Path to the Move package (i.e., where the Move.toml file is)

Implementations

impl CompiledPackage

pub fn get_modules(&self) -> impl Iterator<Item = &CompiledModule>

Return all of the bytecode modules in this package (not including direct or transitive deps) Note: these are not topologically sorted by dependency–use get_dependency_sorted_modules to produce a list of modules suitable for publishing or static analysis

pub fn into_modules(self) -> Vec<CompiledModule>

Return all of the bytecode modules in this package (not including direct or transitive deps) Note: these are not topologically sorted by dependency–use get_dependency_sorted_modules to produce a list of modules suitable for publishing or static analysis

pub fn get_dependent_modules(&self) -> impl Iterator<Item = &CompiledModule>

Return all of the bytecode modules that this package depends on (both directly and transitively) Note: these are not topologically sorted by dependency.

pub fn get_modules_and_deps(&self) -> impl Iterator<Item = &CompiledModule>

Return all of the bytecode modules in this package and the modules of its direct and transitive dependencies. Note: these are not topologically sorted by dependency.

pub fn get_dependency_sorted_modules(&self) -> Vec<CompiledModule>

Return the bytecode modules in this package, topologically sorted in dependency order This is the function to call if you would like to publish or statically analyze the modules

pub fn get_package_bytes(&self) -> Vec<Vec<u8>>

Return a serialized representation of the bytecode modules in this package, topologically sorted in dependency order

pub fn get_package_base64(&self) -> Vec<Base64>

Return the base64-encoded representation of the bytecode modules in this package, topologically sorted in dependency order

pub fn get_framework_modules(&self) -> impl Iterator<Item = &CompiledModule>

Get bytecode modules from the Sui Framework that are used by this package

pub fn get_stdlib_modules(&self) -> impl Iterator<Item = &CompiledModule>

Get bytecode modules from the Move stdlib that are used by this package

pub fn verify_framework_version(
&self,
ext_sui_framework: Vec<CompiledModule>,
ext_move_stdlib: Vec<CompiledModule>
) -> SuiResult<()>

Version of the framework code that the binary used for compilation expects should be the same as version of the framework code bundled as compiled package’s dependency and this function verifies this.

pub fn generate_struct_layouts(&self) -> Registry

Generate layout schemas for all types declared by this package, as well as all struct types passed into entry functions declared by modules in this package (either directly or by reference). These layout schemas can be consumed by clients (e.g., the TypeScript SDK) to enable BCS serialization/deserialization of the package’s objects, tx arguments, and events.

pub fn is_framework(&self) -> bool

Checks whether this package corresponds to a built-in framework

pub fn published_root_module(&self) -> Option<&CompiledModule>

Checks for root modules with non-zero package addresses. Returns an arbitrary one, if one can can be found, otherwise returns None.